README.md
Crimes have been severely increased in past few years, the Problem Statement includes analysis of crimes with different perspectives including utmost attributes possible and predicting via the study of nature of crimes committed. The problem statement is described to initially predict the crime-type based on location and time. We worked on data about historical crimes in California. We had close to 13,000 records of crimes with data on the date and time of the crime, its location, and its type. Common types of crime include theft, criminal damage, criminal trespass, and assault. This project took on the task of predicting the type of crime that was committed given a police report in two ways one according to time that is when crime took place and another is location that is where crime took place. From a small number of overly detailed features, in time it will give the detail that at which time slot which crime is maximum and in location it will tell at which place which type of crime is maximum. They then trained various diagram based models (Graphs and Pie charts) to classify crimes by type using the generated features. Finally, they tested the performance of their models on testing data. They conclude that predicting the type of crimes committed by time and location alone is quite difficult, but that the feature engineering greatly increases predictive power. Predictions will be made to provide local authorities with an upper hand on crime and help them plan a better strategy to tackle the same.